+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [Unreleased]
+
+### Removed
+
+  * Remove the entirety of metadata.ts. Switch to @metaplex-foundation/mpl-token-metadata
+
+### Changed
+
+  * Upgrade to metaplex metadata v2
+
 # [2.2.0](https://github.com/ChewingGlassFund/wumbo-programs/compare/v2.1.11...v2.2.0) (2022-01-25)
 
 ### Added